refactor: changed export location to make imports easier
