+++ src/features/register-player/player-profile/FormPlayerProfile.tsx
@@ -9,8 +9,7 @@ import { Form, FormControl, FormField, FormItem } from '@/components/ui/form'
 import { cn } from '@/lib/utils'
 
 import FormText from '../FormText.json'
-import useSubmitFormRegister from '../utils/submitFormRegister'
-import { FormTitle } from '../utils/title-form'
+import { FormTitle, useSubmitFormRegister } from '../utils'
 import { FormAdditionalText } from './components/FormAdditionalText'
 import { getRpgStyling } from './helpers/getRpgStyle'
 import { useFormPlayerProfile } from './hooks/useFormPlayerProfile'
