+++ src/features/register-player/utils/index.ts
@@ -1,0 +1,4 @@
+export { default as ButtonBack} from './button-back'
+export { getError } from './getError'
+export { default as useSubmitFormRegister } from './submitFormRegister'
+export { FormTitle, FormTitleSkeleton } from './title-form'
